pyproject.toml
[project]
name = "english-idioms-and-words-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.26.0",
    "apscheduler>=3.11.2",
    "groq>=1.1.1",
    "python-dotenv>=1.2.2",
]
